refactor: dto 패키지 제거
