+++ src/main/java/online/bottler/notification/adaptor/in/web/NotificationController.java
@@ -5,9 +5,9 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import online.bottler.global.response.ApiResponse;
-import online.bottler.notification.application.dto.response.NotificationResponse;
-import online.bottler.notification.application.dto.response.SubscriptionResponse;
-import online.bottler.notification.application.dto.response.UnreadNotificationResponse;
+import online.bottler.notification.application.response.NotificationResponse;
+import online.bottler.notification.application.response.SubscriptionResponse;
+import online.bottler.notification.application.response.UnreadNotificationResponse;
 import online.bottler.notification.application.port.NotificationUseCase;
 import online.bottler.notification.application.port.SubscriptionUseCase;
 import online.bottler.notification.domain.NotificationType;
