+++ src/main/java/online/bottler/notification/application/NotificationService.java
@@ -2,9 +2,9 @@ package online.bottler.notification.application;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import online.bottler.notification.application.dto.request.RecommendNotificationCommand;
-import online.bottler.notification.application.dto.response.NotificationResponse;
-import online.bottler.notification.application.dto.response.UnreadNotificationResponse;
+import online.bottler.notification.application.request.RecommendNotificationCommand;
+import online.bottler.notification.application.response.NotificationResponse;
+import online.bottler.notification.application.response.UnreadNotificationResponse;
 import online.bottler.notification.application.port.NotificationUseCase;
 import online.bottler.notification.application.port.NotificationPersistencePort;
 import online.bottler.notification.application.port.SubscriptionPersistencePort;
