+++ src/main/java/online/bottler/notification/application/SubscriptionService.java
@@ -2,7 +2,7 @@ package online.bottler.notification.application;
 
 import lombok.RequiredArgsConstructor;
 import online.bottler.global.exception.ApplicationException;
-import online.bottler.notification.application.dto.response.SubscriptionResponse;
+import online.bottler.notification.application.response.SubscriptionResponse;
 import online.bottler.notification.application.port.SubscriptionUseCase;
 import online.bottler.notification.application.port.SubscriptionPersistencePort;
 import online.bottler.notification.domain.Subscription;
