+++ src/main/java/online/bottler/notification/application/port/SubscriptionUseCase.java
@@ -1,6 +1,6 @@
 package online.bottler.notification.application.port;
 
-import online.bottler.notification.application.dto.response.SubscriptionResponse;
+import online.bottler.notification.application.response.SubscriptionResponse;
 
 public interface SubscriptionUseCase {
