+++ src/main/java/online/bottler/notification/application/request/NotificationLabelRequestDTO.java
@@ -1,4 +1,4 @@
-package online.bottler.notification.application.dto.request;
+package online.bottler.notification.application.request;
 
 public record NotificationLabelRequestDTO(
         Long letterId,
