+++ src/main/java/online/bottler/notification/application/response/UnreadNotificationResponse.java
@@ -1,4 +1,4 @@
-package online.bottler.notification.application.dto.response;
+package online.bottler.notification.application.response;
 
 public record UnreadNotificationResponse(
         long count
